// app.client/src/api/models/EventTypeListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EventTypeListVM = {
    eventTypeId?: string | null;
    name?: string | null;
    code?: string | null;
    description?: string | null;
    colorOrIcon?: string | null;
    isActive?: boolean;
    category?: string | null;
    categoryId?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};